serverless/test2/requirements.txt
bs4
selenium
python-dateutil
lxml
pandas
